requirements.txt
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
filelock==3.17.0
idna==3.10
numpy==2.2.2
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
